[OPS]  Added fallback to using a profile-less key
Problem?

When we deploy in aws, we typically set credentials at the instance
level, and we don't have a profile name, while on dev local, we have aws
profiles configured and ready to go.  But having different profile names
in different environments breaks waddle when we set our meta parameters.

Solution!

default to using a profile-less session for those cases where the
profile does not exist in the list of available profiles.